Move `CallInfo`-related structs to their own file.
