+++ crates/blockifier/src/execution.rs
@@ -1,4 +1,5 @@
 pub mod cairo1_execution;
+pub mod call_info;
 pub mod common_hints;
 pub mod contract_address;
 pub mod contract_class;
