+++ crates/blockifier/src/execution/deprecated_syscalls/deprecated_syscalls_test.rs
@@ -11,7 +11,8 @@ use starknet_api::{calldata, class_hash, contract_address, patricia_key, stark_f
 use test_case::test_case;
 
 use crate::abi::abi_utils::selector_from_name;
-use crate::execution::entry_point::{CallEntryPoint, CallExecution, CallInfo, CallType, Retdata};
+use crate::execution::call_info::{CallExecution, CallInfo, Retdata};
+use crate::execution::entry_point::{CallEntryPoint, CallType};
 use crate::retdata;
 use crate::state::state_api::StateReader;
 use crate::test_utils::{
