+++ crates/blockifier/src/transaction/objects.rs
@@ -6,7 +6,7 @@ use starknet_api::hash::StarkFelt;
 use starknet_api::stark_felt;
 use starknet_api::transaction::{Fee, TransactionHash, TransactionSignature, TransactionVersion};
 
-use crate::execution::entry_point::CallInfo;
+use crate::execution::call_info::CallInfo;
 use crate::transaction::errors::TransactionExecutionError;
 
 pub type TransactionExecutionResult<T> = Result<T, TransactionExecutionError>;
