+++ crates/blockifier/src/transaction/transactions.rs
@@ -8,10 +8,10 @@ use starknet_api::transaction::{
 
 use crate::abi::abi_utils::selector_from_name;
 use crate::block_context::BlockContext;
+use crate::execution::call_info::CallInfo;
 use crate::execution::contract_class::ContractClass;
 use crate::execution::entry_point::{
-    CallEntryPoint, CallInfo, CallType, ConstructorContext, EntryPointExecutionContext,
-    ExecutionResources,
+    CallEntryPoint, CallType, ConstructorContext, EntryPointExecutionContext, ExecutionResources,
 };
 use crate::execution::execution_utils::execute_deployment;
 use crate::state::cached_state::{CachedState, TransactionalState};
