+++ crates/blockifier/src/transaction/transactions_test.rs
@@ -22,10 +22,9 @@ use crate::abi::abi_utils::{
 };
 use crate::abi::constants as abi_constants;
 use crate::block_context::BlockContext;
+use crate::execution::call_info::{CallExecution, CallInfo, OrderedEvent, Retdata};
 use crate::execution::contract_class::{ContractClass, ContractClassV0, ContractClassV1};
-use crate::execution::entry_point::{
-    CallEntryPoint, CallExecution, CallInfo, CallType, OrderedEvent, Retdata,
-};
+use crate::execution::entry_point::{CallEntryPoint, CallType};
 use crate::execution::errors::EntryPointExecutionError;
 use crate::fee::fee_utils::calculate_tx_fee;
 use crate::fee::gas_usage::{calculate_tx_gas_usage, estimate_minimal_fee};
